refactor: add infrastructure extension
